added check if we are root (for manual builds)
git-svn-id: https://svn.fsinf.at/repositories/scripts@522 997389ca-7026-0410-9b2b-f7b85ba8ea54

+++ build_spectrum.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [[ $(whoami) != "root" ]]; then
+	echo "Error: require root privileges."
+	exit 1 
+fi
+
 find /home/mati/build | grep spectrum-dev | xargs rm -f 2> /dev/null
 cd /chroot
 su mati -c "mchroot schroot -c DIR -d /home/mati/repositories/all/spectrum-dev ../../scripts/build.py"
